# contract/openapi/citytrees-public-api-v1.openapi.yml
openapi: 3.0.0
info:
  title: citytrees public API
  contact:
    name: Nagovitsin Maxim
    email: nagovitsin.maxim@gmail.com
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1
    description: local environment
paths:
...
  /tree:
    post:
      tags:
        - tree-controller
      summary: Add new tree
      operationId: createTree
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreeCreateRequest"
      responses:
        200:
          description: Tree successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TreeCreateResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /tree/{id}:
    get:
      tags:
        - tree-controller
      summary: Return tree information
      operationId: getTreeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Tree information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeGetResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - tree-controller
      summary: Update tree
      operationId: updateTreeById
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreeUpdateRequest'
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - tree-controller
      summary: Delete tree by id
      operationId: deleteTree
      security:
        - BearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /tree/{treeId}/approve:
    post:
      tags:
        - tree-controller
      summary: Approve tree
      operationId: approveTree
      security:
        - BearerAuth: [ ]
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /tree/{treeId}/file:
    post:
      tags:
        - tree-controller
      summary: Attach file to tree
      operationId: attachFile
      security:
        - BearerAuth: [ ]
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        200:
          description: File successfully uploaded. It's id is in response's body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /tree/{treeId}/file/all:
    get:
      tags:
        - tree-controller
      summary: Load all attached files
      operationId: getAllAttachedFiles
      parameters:
        - name: treeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Return tree files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreeGetAttachedFileResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /trees:
    get:
      tags:
        - trees-controller
      summary: Load trees in region
      operationId: loadTreesByRegion
      parameters:
        - name: x1
          description: top left latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: y1
          description: top left longitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: x2
          description: bottom right latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: y2
          description: bottom right longitude
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        200:
          description: Trees in specified region
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreesGetResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /trees/clusters:
    get:
      deprecated: true
      tags:
        - trees-controller
      summary: Load trees clusters in region
      operationId: loadTreesClustersByRegion
      parameters:
        - name: x1
          description: top left latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: y1
          description: top left longitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: x2
          description: bottom right latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: y2
          description: bottom right longitude
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        200:
          description: Trees clusters in specified region
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TreesClusterGetResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /tree/all/count:
    get:
      tags:
        - tree-controller
      summary: Count all trees
      operationId: getAllTreesCount
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeCountAllGetResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /file/upload:
    post:
      tags:
        - file-controller
      summary: Uploads file
      operationId: uploadFile
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        200:
          description: File successfully uploaded. It's id is in response's body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /file/{id}:
    get:
      tags:
        - file-controller
      operationId: getFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Returns file's data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileGetResponse'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - file-controller
      operationId: deleteFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /file/download/{id}:
    get:
      tags:
        - file-controller
      operationId: downloadFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Downloads file
          content:
            text/plain:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    TreeStatus:
      type: string
      enum:
        - NEW
        - TO_APPROVE
        - APPROVED
        - DELETED
    TreeState:
      type: string
      enum:
        - ALIVE
        - DEAD
        - STUMP
    TreeCondition:
      type: string
      enum:
        - VERY_BAD
        - BAD
        - NORMAL
        - GREAT
        - AWESOME
    TreeBarkCondition:
      type: string
      enum:
        - HOLLOW
        - GROWTH
        - CRACK
        - MECHANICAL_DAMAGE
    TreeBranchCondition:
      type: string
      enum:
        - DEAD
        - BROKEN
        - CUT
        - DANGEROUSLY_OVERHANGING
    TreePlantingType:
      type: string
      enum:
        - SELF_SEEDING
        - CULTURAL_SEEDING
...
    TreeCreateRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          example: 0.0
        longitude:
          type: number
          format: double
          example: 0.0
    TreeCreateResponse:
      type: object
      required:
        - treeId
      properties:
        treeId:
          type: string
          format: uuid
          example: ed156378-204b-473c-9d2a-886941a819ff
    TreeGetResponse:
      type: object
      required:
        - id
        - userId
        - status
        - latitude
        - longitude
      properties:
        id:
          type: string
          format: uuid
          example: ed156378-204b-473c-9d2a-886941a819ff
        userId:
          type: string
          format: uuid
          example: ed156378-204b-473c-9d2a-886941a819ff
        status:
          $ref: '#/components/schemas/TreeStatus'
        latitude:
          type: number
          format: double
          example: 0.0
        longitude:
          type: number
          format: double
          example: 0.0
        woodTypeId:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/TreeState'
        age:
          type: integer
        condition:
          $ref: '#/components/schemas/TreeCondition'
        barkCondition:
          type: array
          items:
            $ref: '#/components/schemas/TreeBarkCondition'
        branchesCondition:
          type: array
          items:
            $ref: '#/components/schemas/TreeBranchCondition'
        plantingType:
          $ref: '#/components/schemas/TreePlantingType'
        comment:
          type: string
        fileIds:
          type: array
          items:
            type: string
        diameterOfCrown:
          type: number
          format: double
        heightOfTheFirstBranch:
          type: number
          format: double
        numberOfTreeTrunks:
          type: integer
        treeHeight:
          type: number
          format: double
        trunkGirth:
          type: number
          format: double
    TreeGetAttachedFileResponse:
      type: object
      required:
        - id
        - size
        - name
        - url
      properties:
        id:
          type: string
          format: uuid
          example: ed156378-204b-473c-9d2a-886941a819ff
        size:
          type: integer
          format: int64
        name:
          type: string
        url:
          type: string
    TreeUpdateRequest:
      type: object
      required:
        - status
      properties:
        woodTypeId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TreeStatus'
        state:
          $ref: '#/components/schemas/TreeState'
        age:
          type: integer
        condition:
          $ref: '#/components/schemas/TreeCondition'
        barkCondition:
          type: array
          items:
            $ref: '#/components/schemas/TreeBarkCondition'
        branchesCondition:
          type: array
          items:
            $ref: '#/components/schemas/TreeBranchCondition'
        plantingType:
          $ref: '#/components/schemas/TreePlantingType'
        comment:
          type: string
        fileIds:
          type: array
          items:
            type: string
            format: uuid
        diameterOfCrown:
          type: number
          format: double
        heightOfTheFirstBranch:
          type: number
          format: double
        numberOfTreeTrunks:
          type: integer
          format: double
        treeHeight:
          type: number
          format: double
        trunkGirth:
          type: number
          format: double
    TreeCountAllGetResponse:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          format: int64
...
    TreesGetResponse:
      type: array
      items:
        $ref: '#/components/schemas/TreesGetResponseTree'
    TreesGetResponseTree:
      type: object
      required:
        - id
        - latitude
        - longitude
        - status
      properties:
        id:
          type: string
          format: uuid
          example: ed156378-204b-473c-9d2a-886941a819ff
        latitude:
          type: number
          format: double
          example: 0.0
        longitude:
          type: number
          format: double
          example: 0.0
        status:
          $ref: '#/components/schemas/TreeStatus'
    TreesClusterGetResponse:
      type: object
      required:
        - latitude
        - longitude
        - count
      properties:
        latitude:
          type: number
          format: double
          example: 0.0
        longitude:
          type: number
          format: double
          example: 0.0
        count:
          type: integer
          format: int64
...
    FileUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
    FileUploadResponse:
      type: object
      required:
        - fileId
        - url
      properties:
        fileId:
          type: string
          format: uuid
          example: 70c47221-2181-40fb-8909-4d7e3549c4d4
        url:
          type: string
    FileGetResponse:
      type: object
      required:
        - id
        - name
        - mimeType
        - size
        - hash
        - userId
      properties:
        id:
          type: string
          format: uuid
          example: 70c47221-2181-40fb-8909-4d7e3549c4d4
        name:
          type: string
          example: "example.txt"
        mimeType:
          type: string
          example: "text/plain"
        size:
          type: integer
          format: int64
          example: 12
        hash:
          type: string
          example: 674401448c4bc69e39b49a7d867058f4
        userId:
          type: string
          format: uuid
          example: 1a60dc75-6e43-4e62-becf-2683042b4818
    ErrorResponse:
      description: Error
      type: object
      required:
        - code
        - message
      properties:
        code:
          description: Error code, identifier for type of error
          type: string
          example: Forbidden
        message:
          description: Description of error
          type: string
          example: Something went wrong
  responses:
    BadRequest:
      description: Unexpected request content
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not enough privileges
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    CookieAuth:
      type: apiKey
      in: cookie
      name: ct_refresh_token